Fix SDL_gfx build breaking if aclocal-1.10 is installed
The source archive for SDL_gfx has a configure.in with a newer
timestamp than aclocal.m4.  At the same time, the directory "m4"
needed for rerunning aclocal is missing.  The result is that
the Makefile tries to run aclocal-1.10, and if it exists aclocal will
bail out because of the missing directory.  If aclocal-1.10 is _not_
installed, the old aclocal.m4 is used and all is well.

To prevent the build from failing on systems with aclocal-1.10 installed,
hide the dependency on configure.in.